Extract controllers for /users/* routes
